requirements.txt
numpy==1.24.4
Pillow==10.0.0
pydicom==2.4.3
